requirements.txt
lxml==4.6.3
Flask==2.0.1
requests==2.27.1
beautifulsoup4==4.11.1
googlesearch-python==1.3.0
